+++ NEWS.md
@@ -5,7 +5,7 @@
 ### Breaking changes
 
 * Replaced `chunks` with simpler `qenv` class.
-* Replaced `datasets` argument containing `FilteredData` with the new arguments `data` (list of reactive datasets) and `filter_panel_api` (`FilterPanelAPI`).
+* Replaced `datasets` argument containing `FilteredData` with the new arguments `data` (`tdata` object) and `filter_panel_api` (`FilterPanelAPI`).
 
 # teal.goshawk 0.1.14