+++ NEWS.md
@@ -2,6 +2,11 @@
 
 # Examples now use `scda.2022` data instead of `scda.2021`.
 
+### Breaking changes
+
+* Replaced `chunks` with simpler `qenv` class.
+* Replaced `datasets` argument containing `FilteredData` with the new arguments `data` (`tdata` object) and `filter_panel_api` (`FilterPanelAPI`).
+
 # teal.goshawk 0.1.14
 
 ### Enhancements